refactor: move alloc to internal/memory package
