+++ internal/memory/alloc_go1.20.go
@@ -1,7 +1,7 @@
 //go:build !go1.21
 // +build !go1.21
 
-package internal
+package memory
 
 func Alloc(n int) []byte {
 	return make([]byte, n)
